Prefer flat values structure for helm_deploy
In #4354, `failedJobsHistoryLimit` and `backoffLimit` values were made
configurable on a per-job basis.

A nested values structure was used, however `cronjobs` and `jobs` value hashes
were only defined for the uat environment.

This meant [errors were raised] deploying to staging (or production).

This change prefers a flat structure for these variables.

[errors were raised]: https://app.circleci.com/pipelines/github/ministryofjustice/laa-apply-for-legal-aid/17953/workflows/d4cb9009-ee8c-4383-87b8-2f6831382c33/jobs/94136