Refresh our SimpleCov config
This makes some small (but hopefully useful) changes to our SimpleCov
configuration.

1. Use SimpleCov for Cucumber feature tests

In the absence of traditional Rails system (feature) tests, we use Cucumber to
write feature level integration tests. Until now, though, our coverage reports
have not taken into account the code covered by these important tests. This
seemed like an oversight, so it has been added.

2. Simplify coverage configuration

Now that SimpleCov was being invoked in several places, it made sense to
centralise configuration in a `.simplecov` file as [recommended by the
documentation].

This simplification means we can inline the work to process coverage in the task
itself.

[recommended by the documentation]: https://github.com/simplecov-ruby/simplecov#using-simplecov-for-centralized-config

3. Add branch coverage

Branch coverage is useful because it can highlight untested conditional logic in
the codebase. Currently there is no minimum level (we should do some work to
address this before setting it), however no drop in coverage is allowed.