Fix flaky `ProviderDetailsCreator` spec
This test file contained a [flaky spec].

Trying to isolate the flaky spec in order to reproduce and fix the issue
proved difficult, so I took a more pragmatic approach to resolve the issue.

1. Ensure each test is self-contained
2. Ensure each test is repeatable (i.e deterministic)
3. Avoid using `let!` (linked to step 1.)

In the process of refactoring the tests, it also highlighted some tests that
didn't add any value -those have been removed.

I'm confident the refactored tests are not flaky, and will improve the
robustness of our test suite -saving us frustration, time, and money.

[flaky spec]: https://app.circleci.com/pipelines/github/ministryofjustice/laa-apply-for-legal-aid/18165/workflows/ae431d1e-b060-4638-8b78-91df79353cc1/jobs/95186/steps