Do not display inapplicable transaction types on the CYA page
Before, "Disregarded benefits" was displayed in the income section of the
check your answers page for all applications.

This did not make sense for applications that are not using open banking,
though.

Now, the "Disregarded benefits" section is only displayed for applications using
open banking and the "Housing benefits" section is never displayed.

When applications are uploading bank statements, neither the "Disregarded
benefits" or "Housing benefits" sections are displayed.

Cucumber matchers for the check your answers pages have also been updated to be
more specific and expect given values in the income section.