Amend axios handling
Amend @babel/plugin-proposal-class-properties to not use loose mode
to remove warning

And import 'axios' rather than require it, but not sure this is
the right thing to do but fixed the problem locally. Previoulsy
this resulted in a CI lint failure as below:

```
FAIL spec/javascript/worker_waiter.spec.js
  ● worker is complete after 2 API calls › polls the correct endpoint twice

    TypeError: Cannot read properties of undefined (reading 'get')

      3 | async function checkWorkerStatus () {
      4 |   const workerId = document.querySelector('.worker-waiter').getAttribute('data-worker-id')
    > 5 |   const response = await axios.get(`/v1/workers/${workerId}`)
        |                                ^
      6 |   return response.data
      7 | }
      8 |

      at Object.get [as checkWorkerStatus] (app/webpack/src/worker_waiter.js:5:32)
      at Object.checkWorkerStatus (spec/javascript/worker_waiter.spec.js:36:18)

Though the "loose" option was set to "false" in your @babel/preset-env config, it will not be used for @babel/plugin-proposal-private-property-in-object since the "loose" mode option was set to "true" for @babel/plugin-proposal-class-properties.
The "loose" option must be the same for @babel/plugin-proposal-class-properties, @babel/plugin-proposal-private-methods and @babel/plugin-proposal-private-property-in-object (when they are enabled): you can silence this warning by explicitly adding
  ["@babel/plugin-proposal-private-property-in-object", { "loose": true }]
to the "plugins" section of your Babel config.
PASS spec/javascript/modules/proceedings.spec.js
```