Prevent basic uploads navigating to truelayer `income_summary`
This is live bug whereby if the provider does not add any
outgoings on the outgoings page but has provided incomes
on the earlier incomes page they will be take to the
income_summary page - this page is only for categorising
bank transactions and part of the truelayer path only.

ps: this would not affect the enhanced journey.