Remove housing benefit from income section
The income section of the review and print page
is only to be used by truelayer applications
for which the housing benefit transaction type
is not applicable.

The enhanced bank statement journey will exclude this
entire section will sho housing benefit in its own
section according to designs and will not show
the entire "income, regular payments and assets" section
in any event.