Refactor regular transaction forms (#4404)
* Refactor regular transaction forms

This introduces a new `RegularTransactionForm` model for persisting regular
transaction data for an application (and destroying dependent records).

Each of `RegularIncome`, `RegularOutgoings` and `HousingBenefit` forms make
use of this new abstract form.

The basic premise of this refactor is that each of the forms exhibit the same
characteristic behaviour:

1. They validate user input (i.e an option is selected, inputs are valid).

2. They facilitate validation and creation of regular transaction records that
   belong to a given legal aid application.

3. They persist some data on the legal aid application record itself.

4. They destroy the correct combination of associated records. This is currently
   complicated by several factors:
    - The relationship between LegalAidApplication <> TransactionType <>
      Transaction (regular/cash) is indirect.
    - Housing payments (rent_or_mortgage outgoings) are coupled to housing
      benefit. This means housing benefit transactions can only exist if there is
      a rent or mortgage transaction. This also impacts user flow.
    - Housing benefit is also an income (i.e credit) transaction (although it is
      a child transaction type), but it is not handled by the income form and
      income form changes should not affect housing benefit transactions.

5. They assign regular transaction data from the legal aid application if it
   exists (i.e so correct data is displayed when changing answers).

Each form subclass defines a `transaction_type_conditions` private method that
sets out the conditions with which associated transactions should be scoped.
This means the `RegularIncomeForm` is only interested in transactions where the
transaction type has a "credit" operation and "nil" parent_id, for example.

Each form subclass also defines a `legal_aid_application_attributes` private
method that outlines the attributes to be updated for a given form.

This change also highlights further areas for refactoring -in particular the
relationship between a legal aid application, transaction types, and types of
transaction (i.e regular, cash, bank).

Subsequent work will be done to explore the feasibility of simplifying those
relationships. Essentially all types of transaction should be joined to an
application through the existing `LegalAidApplicationTransactionType` join
table.

* Update housing benefit form error messages

Before, the blank error message was indirect.

This updates the error message to be clearer and more instructive.

All error messages are updated to match the content tweak in #4450 that
capitalised `Housing Benefit`.