Send applicant employed status to CFE
In order for CFE to apportion child care cost allowance for an applicant, it
needs to know whether an applicant is employed or not.

This adds `employed` to the `CreateApplicantService` payload, so that the
data point can be persisted to the applicant in CFE.

[Changes have been made to CFE][cfe-changes] to take into account this new
information when collating the childcare costs allowance for an assessment.

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3560)

[cfe-changes]: https://github.com/ministryofjustice/check-financial-eligibility/pull/1163