Providers can categorise disregarded benefit transactions
This behaviour was erroneously removed in #4366.

Once a client has provided access to their transactions using TrueLayer,
providers need to be able to categorise their bank transactions.

One of these categories is "Disregarded benefits".

In #4366, the `TransactionType.find_with_children` method was removed (alongside
its use in `Providers::Means::IdentifyTypesOfIncomesController`.

What we didn't know then (but we do now) is that the intention of this method
was to allow child transaction types to be associated to an application when
their parent transaction type was identified.

For instance:

  When "Benefits" are released, we want to create "Disregarded benefits"
  transaction types too.

This updates the affected controller to create transaction type records for all
identified credit transaction types including their children (but not housing
benefit).

A feature test has been added, since this behaviour was previously untested
--hence the regression!