Add matter opposition question to merits task list
When we ask the provider to provide information regarding the merit of the
application for legal aid, we ask them to complete the merits task list.

The merits task list is made up of sets of questions that apply to either the
application as a whole, or to the individual proceedings that make up the
application.

This adds a new application-level question to the merits task list.

The new question is applicable for all client-involvement types and is relevant
for any applications with at least one section 8 proceeding.

This will allow us to capture information from providers about why a given
matter is opposed by either the client or the opposing party (depending on
involvement type) in a free-text input.i

The questions in the merits task list are retrieved from the Legal Framework
API, and this adds a new flow step to the merits task loop to handle the
`why_matter_opposed` task.

When providers submit the form, matter opposed reasons are added to matter
opposition records. These are associated to the application via a `has_one`
relationship, with the most recent record taking precedence.

Typically, an application will only ever have at most one matter opposition
(i.e existing records are updated), but this is not a strict rule —and as such
there are neither unique validations on the model, nor a unique constraint on
the database table.

The existing View <> Controller <> FormModel <> Model pattern in use across the
application has been adopted.

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3490)