Amend check your answers for yes/no national insurance number question
Design and content specifies showing NINO or "not provided" as
the answer to the question.

In addition we conditionally show or hide the "What happens next"
section since it is only applicable to having a NiNO and we
have an interrupt page to warn the user on the next page in any event
(if they do not provide a nino).

Feature test the various flows from this page, including the national
insurance number amendment path.

NOTE: the check you answers flow jump from details to nino question
This is because of a flow where a failed DWP check offers
the user the oppertunity to amend the details. This flow
should including the national insurance number. Separating the
flow logic into two:

1. for amending the details after a failed DWP check
2. for the common CYA page back and forth

needs separate consideration IMO and is not simple to
change.