Add separate migration to backfill has_national_insurance_number
An existing application with a value for
`applicant.national_insurance_number` should, for data
consistency, have has_national_insurance_number
set to true.

Worth noting however that we are not, as yet,
relying on `has_national_insurance_number?` for
flows - instead we can simply rely on national_insurance_number?

There are 30k+ records to update