Fix flaky bank statement spec
Before, this test was flaky because it would try to navigate to the next page
while a document was being uploaded.

In #4375, that document upload spec expected to see "document.pdf UPLOADED"
content before clicking "Save and continue". This ensured that the document had
finished uploaded before navigating away.

This test appeared to work similarly, except it would fail sporadically.

On closer inspection, it turns out the "hello_world.pdf" file was already
uploaded. This meant the test attempted to navigate away by clicking "Save and
continue" straight away (rather than waiting for the upload to finish).

This updates the test to upload a different second file, and expect both files
to be uploaded before navigating away from the page.

As in #4375, similar considerations should be made, but in the meantime this
prevents the test from being flaky.