Fix student finance irregular income bug
Before, student loan irregular income records were created every time the
`StudentFinances::AnnualAmountForm` was saved.

This meant when a provider specified that their client *did not* receive student
finance, an irregular income record with an amound of `nil` was persisted.

Since the boolean answer to this "Yes/No" question is stored on the legal aid
application already, there is no need to create an irregular income record that
captures the absence of data.

The existing implentation was confusing, and apparently unintentional. If we
consider the "Review and print your application" page, student loan data is
rendered in the "Income, payments, and assets" section if there is *any*
irregular income data. This implied that irregular income data should only exist
if the client is in receipt of student finance.

This change resolves that bug by ensuring that student loan irregular income
records are only persisted when the client is in receipt of student finance, and
they have specified an amount.

A separate bug relating to "orphaned" irregular income records has also been
fixed.

Before, if a provider selected "No" after previously selecting "Yes" and
specifying how much, the irregular income record was never updated.

This resolves that issue by destroying associated student loan irregular income
records when the client does not receive student finance.