Communicate that there is no student finance in review
Before, when a provider reviewed or printed an application, they would see an
empty "Student loan" row in the income table if they had specified that the
client did not receive any student finance.

This was due to a bug involving erroneous irregular income records being created
when answering "No" to the student finance question.

Now that issue has been resolved, there is no indication to the provider whether
or not the applicant was in receipt of student finance or not when they review
or print the application.

This updates the review or print page to clearly communicate when the provider
has told us that the client does not receive student finance.

Now, a section has been added (similarly to the non-TrueLayer page) that
displays

```
Does your client receive student finance?             No
```

Existing behaviour (i.e a "Student loan" row in the income table) persists for
applications where the client is in receipt of student finance.