Replace multiline anchors with start and end of string anchors
Nino validation uses the multiline anchors but should
use start and end instead. I noticed this raises error
when spiking use of `validate :blah, :format: { with: regex }`
rails valiadaion

```
The provided regular expression is using multiline anchors (^ or $), which may present a security risk. Did you mean to use \A and \z, or forgot to add the :multiline => true option?
```