Update `govuk-error-summary` usage
In [GOVUK Frontend
4.4.0](https://github.com/alphagov/govuk-frontend/releases/tag/v4.4.0), the HTML for the error summary was updated.

The error summary is typically rendered by the form builder (which has updated
the HTML), but in some instances we are rendering the error summary outside the
form builder.

In these scenarios, the HTML as been updated in line with the recommended
changes.

1. Remove `aria-labelledby="error-summary-title"` and `role="alert"` from the
parent element (`.govuk-error-summary`).
2. Add a div wrapper around the contents of `govuk-error-summary` with the
attribute `role="alert"`.
3. Remove `id="error-summary-title"` from the error summary `h2` element.