Update schema.rb to reflect nullable columns
These columns were added as nullable string columns in #4048, but for some
reason the schema did not reflect this until #4530.

Since then, #4529 undid the change in #4530, so this redoes the change again.

Before, the columns were nullable but specified as `null: false` in the schema.

Now, they are still nullable, and the schema reflects that.