Update schema.rb to reflect production
1. Timestamp precision

This was originally addressed in 22fe6852d095d7d096d1de790af58413e0187bd0 but
the changes were since merged in a different change inadvertently.

Since Rails 7.0, datetime column precision is set to 6 by default.

This means `t.datetime "partied_at"` would generate a column in postgres of
`timestamp(6) without time zone` type.

Before Rails 7.0, datatime column precision was set to nil by default.

This meant `t.datetime "partied_at"` would have generated a column in postgres
of `timestamp without time zone` type.

It seems that we always used the default, which means any datetime columns added
to the database **before** Rails 7.0 have a type without any precision, but any
datetime columns added to the database **after** Rails 7.0 have a type with
precision (6).

Currently our schema suggests that all datetime columns in the database have
precision (6), but this isn't the case.

This updates the datetime columns in the schema to reflect that some columns
have a type without precision (nil) while others have a type with precision (6).

2. Nullable columns

This was originally addressed in ebdf9e8522ab65cbee8881ab0a2fed6131ebe452 but
the changes were since merged in a different change inadvertently.

These columns were added as nullable string columns in #4048, but for some
reason the schema did not reflect this until #4530.

Since then, #4529 undid the change in #4530, so this redoes the change again.

Before, the columns were nullable but specified as `null: false` in the schema.

Now, they are still nullable, and the schema reflects that.

3. Column ordering

Less significant, but discrepancies between production and the schema make
conflicts more likely and make it harder to ensure the correct changes are added
to the schema after running migrations.

4. Misnamed index

The index is named unconventionally in the schema. It has been updated to
reflect the naming in production.