Address Rubocop erblint failures in app/views/shared
Before, all templates in `app/views/shared` were excluded from
Rubocop linting.

This removes that rule (except for `check_answers` and `forms` folders), and
resolves all failures.

All failures were resolved with the `--autocorrect` flag, and are all trivial

One caveat is that when assigning local variables in partials inside a multiline
block, Rubocop thinks the first line is empty. This is a limitation of using
Rubocop in erblint, since each line is parsed in isolation.