Fix: Update ScheduledMailingsDeliveryJob
Currently the scheduled mailer runs each morning on each
UAT instance.  If K8s has deleted the postgres pod sidekiq
will try multiple times to connect and send mails.  This
results in thousands of sentry errors

This PR attempts to allow the job to fail politely.  If
it cannot find the DB it will send a single Sentry message
and then allow the job to exit, this should mean that sidekiq
won't keep retrying as the job jas not 'failed'