Add #non_passported? spec
This was not explicitly tested as it is a simple
invers of passported but no looking to introdouce
the notion of `#non_means_tested?` and to exclude
these from `#non_passported?` logic.

As a result it would seem sensible to apply regression
unit tests as a first step since this logic is used in
quite a few palces in the app.