Amend #non_passported? logic to exclude #non_means_tested?
We now have three possible journeys coming out of the
applicant and proceeding type section's answers. These are:

- passported:
  benefit checked and applicant in receipt of passported benefits

- non passported:
  benefit checked and applicant not in receipt of passported benefits
  BUT NOT non means tested

- non means tested
  benefit check skipped because MTR phase 1 enabled and applicant under 18

Currently the MTR phase one under 18 non means tested situation is "derived"
from the fact that the state machine has been set to "NonMeansTestedStateMachine"
which, in the full running of the application, is set ONLY when MTR phase 1 is
enabled and the applicant is determined to be under 18 at the point just prior
to benefit checking (which is skipped).