Refactor `uploading_bank_statements?` method
Before, this method relied on the private `client_open_banking_consent?`
predicate method.

This method returned `true` if the `provider_received_citizen_consent` attribute
on the application was `nil`, or `provider_recevied_citizen_consent?` otherwise.

This private predicate was then negated in the `uploading_bank_statements?`
method to determine whether an application should use the bank statement upload
journey.

While effective, the intention was slightly unclear and confusing.

This renames and refactors the method to add clarity.