Handle duplicate mail sending and refactor
Multiple, upto 4/5, emails were being sent to
the apply team inbox and, in this PR, to the
provider, for undeliverable emails.

This appears to have been the result of a multiple
EmailMonitorJob scheduling causing multiple processes
to handle status updates and permanent-failures.

This has been fixed by not having the `EmailMonitorJob`
reschedule itself. It is already enqueued by the
`ScheduledMailingsDeliveryJob` which, in turn is
scheduled when various emails are sent and additionally
by a cronjob (see helm_deploy/apply-for-legal-aid/templates/cronjob-deliver-scheduled-mail.yaml) that runs every 30 minutes.