Modify applicant#age to apply non means tested logic
For non means tested (NMT) applications the age should
be based on the value stored in the db. This is set
at the relevant point which for NMT is when providers
continue past the first check your answers page.

Also add this adds a non means tested means report
feature test scenario. However, the means report may
need to NOT be generated for NMT applications but for
now this tests that it can be generated as is.