Generate PDF reports from HTML using `grover` gem
Before, the [`wicked_pdf` gem][wicked] was used to generate PDF files for the
means and merits reports from HTML templates. This gem relies on
[`wkhtmltopdf`][wk] to generate PDFs from HTML.

As part of the [work to upgrade to Ruby version to 3.1.3][ruby], we discovered that
`wkhtmltopdf` was no longer available as part the Alpine image.

As outlined in the repo, the project is no longer actively maintained —so we
took a decide to migrate away from that tool and consider alternatives.

The [`grover` gem][grover] makes use of [`puppeteer`][puppeteer] to render HTML
in a headless browser and save a PDF file. Other approaches were identified (e.g
writing PDFs [using `prawn`][prawn]), but ultimately `grover` was chosen as a
drop-in replacement.

This replaces `wicked_pdf` with `grover` for means and merits report creation
(which are subsequently sent to CCMS), and updates the respective controllers so
that PDF files can be rendered in the browser for testing purposes.

[wicked]: https://github.com/mileszs/wicked_pdf
[wk]: https://github.com/wkhtmltopdf/wkhtmltopdf
[ruby]: https://github.com/ministryofjustice/laa-apply-for-legal-aid/pull/4663#issuecomment-1340628604
[grover]: https://github.com/Studiosity/grover
[puppeteer]: https://github.com/puppeteer/puppeteer
[prawn]: https://github.com/prawnpdf/prawn