Remove `regexp-examples` gem
Before, the [`regexp-examples`][regexp] gem was used to generate application
reference numbers.

The reference numbers we need to generate are trivial, so we can generate the
number by sampling the character set instead.

This removes the `regexp-examples` gem, and reimplements the private method
`#random_reference_number` to ensure it returns valid, random reference numbers.

The tests have been refactored to test the uniqueness and randomness of
reference number generation in a deterministic way without stubbing, by
leveraging `Kernel.srand` to seed random number generation.

The public API has also been refactored slightly to `ReferenceNumber.generate`,
which is a deviation from the strict "service-object" pattern, but provides a
more readable interface for users to interact with.

Reference number generation still utilises a loop to cycle through random
numbers to find a unique one. This will become more expensive over time, but
there are 427,518,000 permutations —so we should be ok for a reasonably
forever-shaped length of time.

[regexp]: https://github.com/tom-lord/regexp-examples

Unit tests are somewhat tied to the implementation in that they stub calls to
`LegalAidApplication.find_by(application_ref:)`. This is not ideal, but it
removes a source of flakiness that would currently prove difficult to unpick.

When creating a `LegalAidApplication` record with a given reference number, the
factory subequently makes lots of calls to `Faker` which results in lots of
calls to `.rand`. This means the generated reference number is much further
ahead in the sequence. This also means, the test is prone to being flaky, since
any addition/removal of "fake" attributes in the legal aid application or
associated factories would change the reference number that was generated.

The current approach avoids this by simply not creating any records. That way,
the sequence of random numbers generated is deterministic and corresponds
directly with the behaviour you would see outside the test (for example calling
`Kernel.srand(1234)` and `ReferenceNumber.generate` several times in the
terminal.

This approach (and the journey to get there) is outlined in more detail in this
[comment on the pull request][comment].

[comment]: https://github.com/ministryofjustice/laa-apply-for-legal-aid/pull/4707#discussion_r1050542830