Do not generate blocks for those not generated for passported
Most of these do appear in the noon means tested example
provided by the BA but I do not believe they are required
for successfull injection since they are ommitted from
passported.

The attributes in question appear to mainly be means related
and therefore not applicable to NMT applications. Others
 are declaration related which are not really revelant
 but we do not know whether they are required by CCMS or
 PUI.