Update Ruby to v3.1.3
Ruby 3.1.3 was [released on 24 November 2022][ruby].

This updates the Ruby version to this latest patch, although the upgrade path is
not straightforward.

There is no `ruby:3.1.3-alpine3.14` docker image, and `wkhtmltopdf` has been
[removed from alpine 3.15 and beyond][alpine]. We use `wicked_pdf` to generate
means and merits PDF reports, and the gem depends on `wkhtmltopdf`.

It seems the `wkhtmltopdf` tool is [no longer actively maintained][wkhtmltopdf],
and alternatives are recommended.

An alternative approach using the `grover` gem and puppeteer/chromium to
generate PDF reports from HTML was explored in #4700 and removes `wkhtmltopdf`
as a dependency.

The CircleCI Ruby browsers image takes the node version from v16.14.2 to
v18.12.1, but unfortunately [webpacker does not work on node17+][node17].

The `--openssl-legacy-provider` flag is added to the `NODE_OPTIONS` environment
variable so that webpacker can be used to compile assets on more recent node
versions.

[ruby]: https://www.ruby-lang.org/en/news/2022/11/24/ruby-3-1-3-released/
[alpine]: https://wiki.alpinelinux.org/wiki/Release_Notes_for_Alpine_3.15.0#Significant_removals
[wkhtmltopdf]: https://github.com/wkhtmltopdf/wkhtmltopdf/issues/5160#issuecomment-1010668103
[node17]: https://github.com/rails/webpacker/issues/3235