Add component for the Means report
This wraps the existing provider means report in a ViewComponent and adds
extensive unit tests. These tests make assertions about the rendered content of
the component, and are a direct replacement for the more expensive (and less
extensive) integration tests deleted.

The complexity of the test setup and the volume of tests present is indicative
of a complex and detailed report.

Ideally, the component should be broken down in to more manageable, discrete
components that encapsulate the business logic and render the different sections
(i.e each partial could be replaced with its own component).

This is very much a proof of concept, and open for discussion. I think it
demonstrates how this approach could help tidy up some of the deeply nested
partials, improving testability.

We might decide we don't need to test the means report in this much detail, or
we might decide we should similarly test the mertis report.