Refactor `ConfirmClientDeclaration` form
Before, the `LegalAidApplications::ConfirmClientDeclarationForm` form model had
a custom validation method that checked whether the
`client_declaration_confirmed` attribute was valid.

Since we are only validating whether a checkbox was selected, we can lean on
[Rails built-in validations][validates] instead.

This updates the form to validate that the `client_declaration_confirmed` has
been accepted (i.e the provider has checked the box, so the attribute is `true`
or `"1"`).

[validates]: https://api.rubyonrails.org/classes/ActiveModel/Validations/ClassMethods.html#method-i-validates

The form now also implements a `save` method that handles updating the
`client_declaration_confirmed_at` timestamp on the legal aid application.
Previously this behaviour was implemented in the controller, but that logic is
now encapsulated in the form object.

This simplifies the controller, which now makes use of the
`save_continue_or_draft` helper method.