Ensure client declaration checkbox errors are linked
Before, when a provider tried to proceed with an application without checking
the declaration checkbox, the error in the error summary was not linked to the
form input.

Since this checkbox is the only input in the form, we should link the errors.

This adds the relevant configuration to the form input so the errors are linked.