Means test review declaration page changes
As part of the means test review work, under 18s will be able to apply for legal
aid using the service.

This means that the wording of the client declaration needs to be different when
the applicant is under 18 (i.e a provider has submitted an application that is
not means-tested on their behalf).

This introduces a local `i18n_prefix` variable to the view, which determines
which set of content is rendered when the provider confirming the client
declaration.